Add .zshrc to install script
Add check for `.zshrc` after `.zprofile` to enable installing on a wider linux platform base. Also Make it so that `$PATH` is not modified if `usr/local/bin`lready exists on path in `.zshrc`.

The install script fails if there is only .zshrc but no .zprofile. This is a common setup as .zshrc is actually the place to declare packages etc. for use in any interactive shell and .zprofile is supposed to be only for login shells: [See unix.stackexchange discussion](https://unix.stackexchange.com/questions/71253/what-should-shouldnt-go-in-zshenv-zshrc-zlogin-zprofile-zlogout).